converters/text-recognizer/requirements.txt
fastapi[standard]
faststream[nats]
httpx

conductor-python

easyocr
pyspellchecker
fpdf2

certifi
itsdangerous
Jinja2
MarkupSafe
numpy
opencv-python
pillow
pytesseract